refactor conversion for elgamal
